Handle format warnings (on Windows).
git-svn-id: https://svn.r-project.org/R-packages/trunk@8425 694ef91d-65df-0310-b7bb-92e67a308ead

+++ foreign/src/pfm-read.c
@@ -549,7 +549,7 @@ read_version_data (struct file_handle *h, struct pfm_read_info *inf)
     if (!date)
       return 0;
     if (strlen (date) != 8)
-      lose ((_("Bad date string length %lu"), strlen (date)));
+      lose ((_("Bad date string length %lu"), (unsigned long)strlen (date)));
     if (date[0] == ' ') /* the first field of date can be ' ' in some
 			   windows versions of SPSS */
 	date[0] = '0';
@@ -576,7 +576,7 @@ read_version_data (struct file_handle *h, struct pfm_read_info *inf)
     if (!time)
       return 0;
     if (strlen (time) != 6)
-      lose ((_("Bad time string length %lu"), strlen (time)));
+      lose ((_("Bad time string length %lu"), (unsigned long)strlen (time)));
     if (time[0] == ' ') /* the first field of date can be ' ' in some
 			   windows versions of SPSS */
 	time[0] = '0';
@@ -752,7 +752,7 @@ read_variables (struct file_handle *h)
 	 character set, so we can't check for it. */
       if (strlen ((char *) name) > 8)
 	lose ((_("position %d: Variable name has %lu characters"),
-	       i, strlen ((char *) name)));
+	       i, (unsigned long)strlen ((char *) name)));
       if ((name[0] < 74 /* A */ || name[0] > 125 /* Z */)
 	  && name[0] != 152 /* @ */)
 	lose ((_("position %d: Variable name begins with invalid character"),
