+++ foreign/src/sfm-read.c
@@ -552,7 +552,7 @@ read_machine_int32_info (struct file_handle * h, int size, int count, int *encod
 
   if (size != sizeof (R_int32) || count != 8)
     lose ((_("%s: Bad size (%d or count (%u) field on record type 7, subtype 3.	Expected size %ld, count 8"),
-	h->fn, size, count, sizeof (R_int32)));
+	h->fn, size, count, (long)sizeof (R_int32)));
 
   assertive_bufread(h, data, sizeof data, 0);
   if (ext->reverse_endian)
@@ -619,7 +619,7 @@ read_machine_flt64_info (struct file_handle * h, int size, int count)
 
   if (size != sizeof (R_flt64) || count != 3)
     lose ((_("%s: Bad size (%d) or count (%u) field on record type 7, subtype 4.	Expected size %lu, count 8"),
-	   h->fn, size, count, sizeof (R_flt64)));
+	   h->fn, size, count, (unsigned long)sizeof (R_flt64)));
 
   assertive_bufread(h, data, sizeof data, 0);
   if (ext->reverse_endian)
@@ -800,7 +800,8 @@ read_header (struct file_handle * h, struct sfm_read_info * inf)
   if (hdr.case_size <= 0 || ext->case_size > (INT_MAX
 					      / (int) sizeof (union value) / 2))
     lose ((_("%s: Number of elements per case (%d) is not between 1 and %ld"),
-	   h->fn, hdr.case_size, INT_MAX / sizeof (union value) / 2));
+	   h->fn, hdr.case_size,
+           (long int) (INT_MAX / (int) sizeof (union value) / 2)));
 
   ext->compressed = hdr.compressed;
